pyproject.toml
[tool.poetry]
name = "causal_discovery"
version = "1.0.2"
description = "normal causal discovery server."
authors = ["zhangkun", "zhangdongcheng"]
[[tool.poetry.source]]
name = "hexin"
url = "http://repositories-public.myhexin.com:8081/repository/pypi-public/simple/"
default = true
[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.7.2"
scipy = "^1.5.4"
arrow = "^0.17.0"
pandas = "^1.1.4"
portion = "^2.1.4"
pyfunctional = "^1.4.2"
networkx = "^2.5"
scikit-learn = "^0.23.2"
tqdm = "^4.54.1"
cupy-cuda100 = "^8.4.0"
torch = "^1.7.1"
peon = {url = "http://gitlab.myhexin.com/peon/peon/uploads/cf61769ab07b579d73c5e01c3eb61ab4/peon-0.2.5-py3-none-any.whl"}
typer = "^0.3.2"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
autopep8 = "^1.5.4"
rope = "^0.18.0"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"